Setting resource owner functionality

+++ src/main/java/com/zuehlke/pgadmissions/services/RoleService.java
@@ -1,18 +1,5 @@
 package com.zuehlke.pgadmissions.services;
 
-import static com.zuehlke.pgadmissions.domain.definitions.workflow.PrismRoleTransitionType.CREATE;
-import static com.zuehlke.pgadmissions.domain.definitions.workflow.PrismRoleTransitionType.DELETE;
-
-import java.util.List;
-
-import javax.inject.Inject;
-
-import org.apache.commons.lang.BooleanUtils;
-import org.joda.time.DateTime;
-import org.springframework.context.ApplicationContext;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.google.common.collect.Lists;
 import com.zuehlke.pgadmissions.dao.RoleDAO;
 import com.zuehlke.pgadmissions.domain.comment.Comment;
@@ -27,15 +14,22 @@ import com.zuehlke.pgadmissions.domain.definitions.workflow.PrismScope;
 import com.zuehlke.pgadmissions.domain.resource.Resource;
 import com.zuehlke.pgadmissions.domain.user.User;
 import com.zuehlke.pgadmissions.domain.user.UserRole;
-import com.zuehlke.pgadmissions.domain.workflow.Action;
-import com.zuehlke.pgadmissions.domain.workflow.Role;
-import com.zuehlke.pgadmissions.domain.workflow.RoleTransition;
-import com.zuehlke.pgadmissions.domain.workflow.StateTransition;
-import com.zuehlke.pgadmissions.domain.workflow.WorkflowPropertyConfiguration;
-import com.zuehlke.pgadmissions.domain.workflow.WorkflowPropertyDefinition;
+import com.zuehlke.pgadmissions.domain.workflow.*;
 import com.zuehlke.pgadmissions.exceptions.DeduplicationException;
+import com.zuehlke.pgadmissions.exceptions.PrismForbiddenException;
 import com.zuehlke.pgadmissions.exceptions.WorkflowEngineException;
 import com.zuehlke.pgadmissions.services.helpers.PropertyLoader;
+import org.apache.commons.lang.BooleanUtils;
+import org.joda.time.DateTime;
+import org.springframework.context.ApplicationContext;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import javax.inject.Inject;
+import java.util.List;
+
+import static com.zuehlke.pgadmissions.domain.definitions.workflow.PrismRoleTransitionType.CREATE;
+import static com.zuehlke.pgadmissions.domain.definitions.workflow.PrismRoleTransitionType.DELETE;
 
 @Service
 @Transactional
@@ -83,7 +77,7 @@ public class RoleService {
         return persistentUserRole;
     }
 
-    public void assignUserRoles(Resource resource, User user, PrismRoleTransitionType transitionType, PrismRole... roles) throws Exception {
+    public void assignUserRoles(Resource resource, User user, PrismRoleTransitionType transitionType, PrismRole... roles) {
         if (roles.length > 0) {
             User invoker = userService.getCurrentUser();
             Action action = actionService.getViewEditAction(resource);
@@ -100,6 +94,14 @@ public class RoleService {
             actionService.executeUserAction(resource, action, comment);
             notificationService.sendInvitationNotifications(comment);
         }
+    }
+
+    public void setResourceOwner(Resource resource, User user) {
+        if (getRolesForResource(resource, user).isEmpty()) {
+            throw new PrismForbiddenException("User has no role within given resource");
+        }
+        resource.setUser(user);
+        resourceService.executeUpdate(resource, PrismDisplayPropertyDefinition.valueOf(resource.getResourceScope().name() + "_COMMENT_UPDATED_USER_ROLE"));
     }
 
     public boolean hasUserRole(Resource resource, User user, PrismRoleGroup prismRoles) {
@@ -139,11 +141,11 @@ public class RoleService {
     }
 
     public List<User> getRoleUsers(Resource resource, Role... roles) {
-        return resource == null ? Lists.<User> newArrayList() : roleDAO.getRoleUsers(resource, roles);
+        return resource == null ? Lists.<User>newArrayList() : roleDAO.getRoleUsers(resource, roles);
     }
 
     public List<User> getRoleUsers(Resource resource, PrismRole... prismRoles) {
-        return resource == null ? Lists.<User> newArrayList() : roleDAO.getRoleUsers(resource, prismRoles);
+        return resource == null ? Lists.<User>newArrayList() : roleDAO.getRoleUsers(resource, prismRoles);
     }
 
     public List<User> getRoleUsers(Resource resource, PrismRoleGroup prismRoleGroup) {
@@ -201,10 +203,12 @@ public class RoleService {
     }
 
     public void deleteUserRole(Resource resource, User user, Role role) {
-        UserRole userRole = roleDAO.getUserRole(resource, user, role);
-        validateUserRoleRemoval(resource, userRole.getRole());
+        if(getRolesForResource(resource, user).size() < 2 && resource.getUser().getId() == user.getId()){
+            throw new PrismForbiddenException("Cannot remove the owner");
+        };
+        UserRole userRole = getUserRole(resource, user, role);
         entityService.delete(userRole);
-        reassignResourceOwner(resource);
+        entityService.flush(); // so that getRolesForResource() returns up-to-date values
     }
 
     public List<PrismRole> getRolesByScope(PrismScope prismScope) {
@@ -287,27 +291,8 @@ public class RoleService {
             UserRole userRole = new UserRole().withResource(resource).withUser(user).withRole(role).withAssignedTimestamp(baseline);
             UserRole transitionUserRole = new UserRole().withResource(transitionResource).withUser(user).withRole(transitionRole)
                     .withAssignedTimestamp(baseline);
-    
+
             applicationContext.getBean(roleTransition.getRoleTransitionType().getResolver()).resolve(userRole, transitionUserRole, comment);
-        }
-    }
-
-    private void validateUserRoleRemoval(Resource resource, Role roleToRemove) {
-        Role creatorRole = getCreatorRole(resource);
-        if (creatorRole == roleToRemove) {
-            List<User> creatorRoleAssignments = getRoleUsers(resource, creatorRole);
-            if (creatorRoleAssignments.size() == 1) {
-                throw new Error();
-            }
-        }
-    }
-
-    private void reassignResourceOwner(Resource resource) {
-        User owner = resource.getUser();
-        Role ownerRole = getCreatorRole(resource);
-        if (!hasUserRole(resource, owner, ownerRole.getId())) {
-            User newOwner = getRoleUsers(resource, ownerRole).get(0);
-            resource.setUser(newOwner);
         }
     }
